requirements.txt
mapbox_earcut==2.0.0
numpy==2.3.5
PyOpenGL==3.1.10
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
shiboken6==6.10.1
